requirements.txt
PyPDF2==3.0.1
openai==1.6.1
python-dotenv==1.0.0
pymongo